fix: do not tuck skips inside other regions
When there was a skip from 0 to 2 and another kind of region from 1 to 3, the skip would be narrowed to go from 1 to 2. This meant that a part of a skip that overhang from a subregion would not be subracted from listen time. This was not a good idea for silences, however, as we lost legitimate parts of skips.